AMD64 support: Handle the x86 asm checks differently. AMD64 support: Get the GL pointer types correct. Fix the rest of the valid sign warnings. Fix logging.

+++ trunk/twilight/src/renderer/dyngl.h
@@ -36,6 +36,7 @@
 #define __dyngl_h
 
 #include "SDL_types.h"
+#include <stddef.h>
 
 /* If you need to do this, do it in config.h */
 #ifndef DYNGLENTRY
@@ -141,8 +142,8 @@ typedef double			GLclampd;		/* double precision float in [0,1] */
  * I am 99% sure that nVidia has these types completely WRONG.
  * But this is what they have for x86, at least at the moment.
  */
-typedef int GLintptrARB;
-typedef int GLsizeiptrARB;
+typedef ptrdiff_t GLintptrARB;
+typedef ptrdiff_t GLsizeiptrARB;
 
 #endif /* DYNGL_TYPES */
